Original Attempt
Uploading the first attempt to solve this puzzle! This code is a mess compared with the final solution, I uploaded it for the record and to help others to understand that it is ok to fail. Code comments were added to figure out what was going on and try to solve on the go.